Rename /client to /clients
