+++ pages/clients/login.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import ClientLoginPage from '../../components/client/login';
+import ClientLoginPage from '../../components/clients/login';
 
 const Clients = () => {
     return <ClientLoginPage />;
